requirements.txt
# Flask Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Database ORM
SQLAlchemy==2.0.23

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# XML Parsing (for WeChat messages)
lxml==4.9.3

# LLM API Client
openai==1.3.7  # Compatible with DeepSeek API

# Logging
colorlog==6.8.0

# Date/Time Utilities
python-dateutil==2.8.2

# JSON Schema Validation
jsonschema==4.20.0

# WSGI Server (Production)
gunicorn==21.2.0

# PostgreSQL Driver
psycopg2-binary==2.9.9
